Muita coisa

+++ Repositorio/Repositorio/DAO/Generic/IGenericDAO.cs
@@ -4,7 +4,7 @@ namespace Repositorio.DAO.Generic
 {
     public interface IGenericDAO<T>
     {
-        long Save(T entity);
+        int Save(T entity);
         void Update(T entity);
         void Delete(T entity);
         T Find(int id);
