+++ Repositorio/Repositorio/Entidades/TipoDeArmazenamento.cs
@@ -10,7 +10,7 @@ namespace Repositorio.Entidades
         public virtual string Descricao { get; set; }
         public virtual IList<Armazenamento> Armazenamentos {get;set;}
 
-        public virtual long Adiciona(ISession session)
+        public virtual int Adiciona(ISession session)
         {
             TipoDeArmazenamentoDAO udao = new TipoDeArmazenamentoDAO(session);
             return udao.Save(this);
