+++ Repositorio/Repositorio/Entidades/TipoDeServico.cs
@@ -14,7 +14,7 @@ namespace Repositorio.Entidades
         public virtual string Descricao { get; set; }
         public virtual IList<Pedido> Pedidos { get; set; }
 
-        public virtual long Adiciona(ISession session)
+        public virtual int Adiciona(ISession session)
         {
             TipoDeServicoDAO udao = new TipoDeServicoDAO(session);
             return udao.Save(this);
